refactor: rename subpackage utils into helpers
