+++ shot_fpfh/pipeline.py
@@ -12,6 +12,7 @@ from sklearn.neighbors import KDTree
 from shot_fpfh.analysis import get_incorrect_matches, plot_distance_hists
 from shot_fpfh.core import RigidTransform
 from shot_fpfh.descriptors import ShotMultiprocessor, compute_fpfh_descriptor
+from shot_fpfh.helpers import write_ply
 from shot_fpfh.icp import icp_point_to_plane, icp_point_to_point
 from shot_fpfh.keypoint_selection import (
     select_keypoints_iteratively,
@@ -26,7 +27,6 @@ from shot_fpfh.matching import (
     ransac_on_matches,
     threshold_filter,
 )
-from shot_fpfh.utils import write_ply
 
 
 @dataclass
